+++ modules/Base/Print/Document/Document.php
@@ -4,6 +4,7 @@ abstract class Base_Print_Document_Document
 {
     private $filename;
     private $footer;
+    private $file_extension;
 
     abstract public function document_type_name();
 
@@ -15,6 +16,22 @@ abstract class Base_Print_Document_Document
     public function get_filename()
     {
         return $this->filename;
+    }
+
+    public function set_filename_extension($extension)
+    {
+        $this->file_extension = $extension;
+    }
+
+    public function get_filename_extension()
+    {
+        return $this->file_extension;
+    }
+
+    public function get_filename_with_extension()
+    {
+        $full = $this->get_filename() . '.' . $this->get_filename_extension();
+        return $full;
     }
 
     abstract public function write_text($text);